solucionando problemas de modal
