+++ frontend/src/Css/logUser.css
@@ -1,63 +1,65 @@
-.imgBackgroundLog {
+/* .imgBackgroundLog {
   background-image: url(https://cdn.pixabay.com/photo/2015/09/22/15/42/burger-951896_960_720.jpg);
   background-position: center center;
   background-repeat: no-repeat;
   background-attachment: fixed;
   background-size: cover;
   background-color: #464646;
-}
+} */
 
-.log-form {
-  padding: 423px 0;
-}
-
-.registration-form .form {
-  font-family: 'Oswald', sans-serif;
-
-  background: rgba(0, 0, 0, 0.7);
-  /* background: #bdc3c7; */
-  /* fallback for old browsers */
-  /* background: -webkit-linear-gradient(to right, #2c3e50, #bdc3c7); */
-  /* Chrome 10-25, Safari 5.1-6 */
-  /* background: linear-gradient(to right, #2c3e50, #bdc3c7); */
-  /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
-  max-width: 600px;
-  margin: auto;
-  padding: 50px 70px;
-  border-top-left-radius: 30px;
+/* .log-form {
+  padding: 0px 0;
+   border-top-left-radius: 30px;
   border-top-right-radius: 30px;
   border-bottom-left-radius: 30px;
-  border-bottom-right-radius: 30px;
+  border-bottom-right-radius: 30px; 
+} */
+.modal-content {
+  background: rgba(0,0,0,0.7) !important;
+
+}
+
+.registrationmodal-form .form {
+  font-family: 'Oswald', sans-serif;
+  background: rgba(0,0,0,0.7);
+    
+  /* max-width: 500px; */
+  margin: auto;
+  padding: 50px 70px;
+  /* border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
+  border-bottom-left-radius: 10px;
+  border-bottom-right-radius: 10px; */
   box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.75);
 }
 
-.titulo {
+.titulomodal {
   font-family: 'Oswald', sans-serif;
-  max-width: 768px;
+  /* max-width: 768px; */
   margin: auto;
-  padding: 35px 0;
+  padding: 30px 0;
   text-align: center;
   color: white;
 }
 
-titleAdmFoods {
+.titleAdmFoods {
   font-family: 'Oswald', sans-serif; 
   color: black;
 }
 
-.registration-form .item {
+.registrationmodal-form .item {
   border-radius: 20px;
   margin-bottom: 25px;
   padding: 10px 20px;
 }
 
-.registration-form .create-account {
+.registrationmodal-form .modalBtn {
   font-family: 'Oswald', sans-serif;
   border-radius: 30px;
   padding: 10px 20px;
   font-size: 18px;
   font-weight: bold;
-  background-color: gray;
+background: gray;
   border: none;
   color: white;
   margin-top: 20px;
@@ -72,7 +74,7 @@ titleAdmFoods {
     padding: 50px 30px;
   }
 
-  .registration-form .form-icon {
+  .registrationmodal-form .form-icon {
     width: 70px;
     height: 70px;
     font-size: 30px;
